Updates session set behaviour to consider resets (#41)
Previously, setting a session used an ambiguous 'force' flag to reset
the state of the repository to the requested reference. However, in
certain situations, it is beneficial to know whether the repository
is already at the requested reference and therefore doesn't need to
be reset. To counter this, the 'force' flag was replaced with the
'reset' flag that optionally resets the repository even if it is at
the requested reference.

This change also modifies the session Set function to use a string,
rather than a git.Reference, as the reference to set the repository
to. This decision was made to simplify the interface, and remove
ambiguity on the caller when the caller doesn't know what type of
reference they've been asked to resolve.